PRODUCTS/ContractGenius/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.10.0
requests==2.31.0
